introduce new subpackage raytracing and refactor code accordingly
