+++ artist/core/__init__.py
@@ -1,18 +1,10 @@
 """Bundle all classes responsible for the core functions in ``ARTIST``."""
 
-from artist.core.heliostat_ray_tracer import (
-    DistortionsDataset,
-    HeliostatRayTracer,
-    RestrictedDistributedSampler,
-)
 from artist.core.kinematics_reconstructor import KinematicsReconstructor
 from artist.core.motor_position_optimizer import MotorPositionsOptimizer
 from artist.core.surface_reconstructor import SurfaceReconstructor
 
 __all__ = [
-    "HeliostatRayTracer",
-    "DistortionsDataset",
-    "RestrictedDistributedSampler",
     "KinematicsReconstructor",
     "SurfaceReconstructor",
     "MotorPositionsOptimizer",
