+++ artist/core/kinematics_reconstructor.py
@@ -3,10 +3,10 @@ import pathlib
 from typing import Any, cast
 
 import torch
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 from torch.optim.lr_scheduler import LRScheduler
 
 from artist.core import core_utils, learning_rate_schedulers
-from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.loss_functions import Loss
 from artist.field.heliostat_group import HeliostatGroup
 from artist.io.calibration_parser import CalibrationDataParser
