+++ artist/core/motor_position_optimizer.py
@@ -2,10 +2,10 @@ import logging
 from typing import Any
 
 import torch
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 from torch.optim.lr_scheduler import LRScheduler
 
 from artist.core import learning_rate_schedulers
-from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.loss_functions import FocalSpotLoss, KLDivergenceLoss, Loss
 from artist.field.heliostat_group import HeliostatGroup
 from artist.scenario.scenario import Scenario
