+++ artist/core/surface_reconstructor.py
@@ -3,12 +3,12 @@ import pathlib
 from typing import Any, cast
 
 import torch
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 from torch.optim.lr_scheduler import LRScheduler
 
 import artist.nurbs.utils
 import artist.util.utils
 from artist.core import core_utils, learning_rate_schedulers
-from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.loss_functions import Loss
 from artist.core.regularizers import IdealSurfaceRegularizer, SmoothnessRegularizer
 from artist.field.heliostat_group import HeliostatGroup
