+++ artist/raytracing/blocking.py
