+++ examples/field_optimizations/generate_results.py
@@ -10,8 +10,8 @@ import h5py
 import paint.util.paint_mappings as paint_mappings
 import torch
 import yaml
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 
-from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.kinematics_reconstructor import KinematicsReconstructor
 from artist.core.loss_functions import FocalSpotLoss, KLDivergenceLoss
 from artist.core.motor_position_optimizer import MotorPositionsOptimizer
