+++ examples/paint_plots/flux_prediction_raytracing.py
@@ -9,8 +9,8 @@ import paint.util.paint_mappings as paint_mappings
 import torch
 import yaml
 from PIL import Image
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 
-from artist.core import HeliostatRayTracer
 from artist.io.paint_calibration_parser import PaintCalibrationDataParser
 from artist.io.paint_scenario_parser import extract_paint_heliostat_properties
 from artist.scenario import Scenario
