+++ tutorials/01_single_heliostat_raytracing_tutorial.py
@@ -7,8 +7,8 @@ import h5py
 import matplotlib.pyplot as plt
 import torch
 from matplotlib.pyplot import tight_layout
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 
-from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.scenario.scenario import Scenario
 from artist.util import index_mapping, set_logger_config
 from artist.util.environment_setup import get_device
