+++ tutorials/02_heliostat_raytracing_distributed_tutorial.py
@@ -5,8 +5,8 @@ import pathlib
 import h5py
 import torch
 from matplotlib import pyplot as plt
+from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 
-from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.field.heliostat_group import HeliostatGroup
 from artist.scenario.scenario import Scenario
 from artist.util import config_dictionary, index_mapping, set_logger_config
